Styled recyclerview for create activity. Add find recipe button.
Added the correct background color, rounded corners, and stylized list elements for filters recyclerview. The find recipe button currently only takes user to RecipesListActivity, the code for the api call still needs to be added, but I wasn't sure if this was done in CreateActivity or RecipesListActivity, so it is just a TODO for now.